my_site/requirements.txt
asgiref==3.2.10
Django==3.0.8
django-bootstrap3==14.1.0
django-braces==1.14.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-js-asset==1.2.2
Pillow==7.2.0
python-http-client==3.2.7
pytz==2020.1
sendgrid==6.4.1
six==1.15.0
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
